+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XR/TrackedPoseDriver.cs
@@ -1,83 +1,174 @@
 using System;
+using UnityEngine.InputSystem.Controls;
 using UnityEngine.InputSystem.LowLevel;
 
 namespace UnityEngine.InputSystem.XR
 {
     /// <summary>
-    /// The TrackedPoseDriver component applies the current Pose value of a tracked device to the transform of the GameObject.
-    /// TrackedPoseDriver can track multiple types of devices including XR HMDs, controllers, and remotes.
+    /// The <see cref="TrackedPoseDriver"/> component applies the current pose value of a tracked device
+    /// to the <see cref="Transform"/> of the <see cref="GameObject"/>.
+    /// <see cref="TrackedPoseDriver"/> can track multiple types of devices including XR HMDs, controllers, and remotes.
     /// </summary>
+    /// <remarks>
+    /// For <see cref="positionInput"/> and <see cref="rotationInput"/>, if an action is directly defined
+    /// in the <see cref="InputActionProperty"/>, as opposed to a reference to an action externally defined
+    /// in an <see cref="InputActionAsset"/>, the action will automatically be enabled and disabled by this
+    /// behavior during <see cref="OnEnable"/> and <see cref="OnDisable"/>. The enabled state for actions
+    /// externally defined must be managed externally from this behavior.
+    /// </remarks>
     [Serializable]
-    [AddComponentMenu("XR/Tracked Pose Driver (New Input System)")]
-    public class TrackedPoseDriver : MonoBehaviour
+    [AddComponentMenu("XR/Tracked Pose Driver (Input System)")]
+    public class TrackedPoseDriver : MonoBehaviour, ISerializationCallbackReceiver
     {
+        /// <summary>
+        /// Options for which <see cref="Transform"/> properties to update.
+        /// </summary>
+        /// <seealso cref="trackingType"/>
         public enum TrackingType
         {
+            /// <summary>
+            /// Update both rotation and position.
+            /// </summary>
             RotationAndPosition,
-            RotationOnly,
-            PositionOnly
-        }
 
+            /// <summary>
+            /// Update rotation only.
+            /// </summary>
+            RotationOnly,
+
+            /// <summary>
+            /// Update position only.
+            /// </summary>
+            PositionOnly,
+        }
 
         [SerializeField]
         TrackingType m_TrackingType;
         /// <summary>
-        /// The tracking type being used by the tracked pose driver
+        /// The tracking type being used by the Tracked Pose Driver
+        /// to control which <see cref="Transform"/> properties to update.
         /// </summary>
+        /// <seealso cref="TrackingType"/>
         public TrackingType trackingType
         {
-            get { return m_TrackingType; }
-            set { m_TrackingType = value; }
+            get => m_TrackingType;
+            set => m_TrackingType = value;
         }
 
+        /// <summary>
+        /// Options for which phases of the player loop will update <see cref="Transform"/> properties.
+        /// </summary>
+        /// <seealso cref="updateType"/>
         public enum UpdateType
         {
+            /// <summary>
+            /// Update after the Input System has completed an update and right before rendering.
+            /// </summary>
+            /// <seealso cref="InputUpdateType.Dynamic"/>
+            /// <seealso cref="InputUpdateType.BeforeRender"/>
             UpdateAndBeforeRender,
+
+            /// <summary>
+            /// Update after the Input System has completed an update.
+            /// </summary>
+            /// <seealso cref="InputUpdateType.Dynamic"/>
             Update,
+
+            /// <summary>
+            /// Update right before rendering.
+            /// </summary>
+            /// <seealso cref="InputUpdateType.BeforeRender"/>
             BeforeRender,
         }
 
         [SerializeField]
         UpdateType m_UpdateType = UpdateType.UpdateAndBeforeRender;
         /// <summary>
-        /// The update type being used by the tracked pose driver
+        /// The update type being used by the Tracked Pose Driver
+        /// to control which phases of the player loop will update <see cref="Transform"/> properties.
         /// </summary>
+        /// <seealso cref="UpdateType"/>
         public UpdateType updateType
         {
-            get { return m_UpdateType; }
-            set { m_UpdateType = value; }
+            get => m_UpdateType;
+            set => m_UpdateType = value;
         }
 
-        [SerializeField]
+        // Disable warnings that these fields are never assigned to. They are set during Unity deserialization and migrated.
+        // ReSharper disable UnassignedField.Local
+#pragma warning disable 0649
+        [SerializeField, HideInInspector]
         InputAction m_PositionAction;
+        [Obsolete("positionAction has been deprecated. Use positionInput instead.")]
         public InputAction positionAction
         {
-            get { return m_PositionAction; }
+            get => m_PositionInput.action;
+            set => positionInput = new InputActionProperty(value);
+        }
+
+        [SerializeField, HideInInspector]
+        InputAction m_RotationAction;
+        [Obsolete("rotationAction has been deprecated. Use rotationInput instead.")]
+        public InputAction rotationAction
+        {
+            get => m_RotationInput.action;
+            set => rotationInput = new InputActionProperty(value);
+        }
+#pragma warning restore 0649
+        // ReSharper restore UnassignedField.Local
+
+        [SerializeField]
+        InputActionProperty m_PositionInput;
+        /// <summary>
+        /// The action to read the position value of a tracked device.
+        /// Must support reading a value of type <see cref="Vector3"/>.
+        /// </summary>
+        public InputActionProperty positionInput
+        {
+            get => m_PositionInput;
             set
             {
-                UnbindPosition();
-                m_PositionAction = value;
-                BindActions();
+                if (Application.isPlaying)
+                    UnbindPosition();
+
+                m_PositionInput = value;
+
+                if (Application.isPlaying && isActiveAndEnabled)
+                    BindPosition();
             }
         }
 
         [SerializeField]
-        InputAction m_RotationAction;
-        public InputAction rotationAction
+        InputActionProperty m_RotationInput;
+        /// <summary>
+        /// The action to read the rotation value of a tracked device.
+        /// Must support reading a value of type <see cref="Quaternion"/>.
+        /// </summary>
+        public InputActionProperty rotationInput
         {
-            get { return m_RotationAction; }
+            get => m_RotationInput;
             set
             {
-                UnbindRotation();
-                m_RotationAction = value;
-                BindActions();
+                if (Application.isPlaying)
+                    UnbindRotation();
+
+                m_RotationInput = value;
+
+                if (Application.isPlaying && isActiveAndEnabled)
+                    BindRotation();
             }
         }
+
+        /// <summary>
+        /// Stores whether the fields of type <see cref="InputAction"/> have been migrated to fields of type <see cref="InputActionProperty"/>.
+        /// </summary>
+        [SerializeField, HideInInspector]
+        bool m_HasMigratedActions;
 
         Vector3 m_CurrentPosition = Vector3.zero;
         Quaternion m_CurrentRotation = Quaternion.identity;
-        bool m_RotationBound = false;
-        bool m_PositionBound = false;
+        bool m_RotationBound;
+        bool m_PositionBound;
 
         void BindActions()
         {
@@ -87,24 +178,36 @@ namespace UnityEngine.InputSystem.XR
 
         void BindPosition()
         {
-            if (!m_PositionBound && m_PositionAction != null)
-            {
-                m_PositionAction.Rename($"{gameObject.name} - TPD - Position");
-                m_PositionAction.performed += OnPositionUpdate;
-                m_PositionBound = true;
-                m_PositionAction.Enable();
-            }
+            if (m_PositionBound)
+                return;
+
+            var action = m_PositionInput.action;
+            if (action == null)
+                return;
+
+            action.Rename($"{gameObject.name} - TPD - Position");
+            action.performed += OnPositionPerformed;
+            m_PositionBound = true;
+
+            if (m_PositionInput.reference == null)
+                action.Enable();
         }
 
         void BindRotation()
         {
-            if (!m_RotationBound && m_RotationAction != null)
-            {
-                m_RotationAction.Rename($"{gameObject.name} - TPD - Rotation");
-                m_RotationAction.performed += OnRotationUpdate;
-                m_RotationBound = true;
-                m_RotationAction.Enable();
-            }
+            if (m_RotationBound)
+                return;
+
+            var action = m_RotationInput.action;
+            if (action == null)
+                return;
+
+            action.Rename($"{gameObject.name} - TPD - Rotation");
+            action.performed += OnRotationPerformed;
+            m_RotationBound = true;
+
+            if (m_RotationInput.reference == null)
+                action.Enable();
         }
 
         void UnbindActions()
@@ -115,36 +218,51 @@ namespace UnityEngine.InputSystem.XR
 
         void UnbindPosition()
         {
-            if (m_PositionAction != null && m_PositionBound)
-            {
-                m_PositionAction.Disable();
-                m_PositionAction.performed -= OnPositionUpdate;
-                m_PositionBound = false;
-            }
+            if (!m_PositionBound)
+                return;
+
+            var action = m_PositionInput.action;
+            if (action == null)
+                return;
+
+            if (m_PositionInput.reference == null)
+                action.Disable();
+
+            action.performed -= OnPositionPerformed;
+            m_PositionBound = false;
         }
 
         void UnbindRotation()
         {
-            if (m_RotationAction != null && m_RotationBound)
-            {
-                m_RotationAction.Disable();
-                m_RotationAction.performed -= OnRotationUpdate;
-                m_RotationBound = false;
-            }
+            if (!m_RotationBound)
+                return;
+
+            var action = m_RotationInput.action;
+            if (action == null)
+                return;
+
+            if (m_RotationInput.reference == null)
+                action.Disable();
+
+            action.performed -= OnRotationPerformed;
+            m_RotationBound = false;
         }
 
-        void OnPositionUpdate(InputAction.CallbackContext context)
+        void OnPositionPerformed(InputAction.CallbackContext context)
         {
-            Debug.Assert(m_PositionBound);
+            Debug.Assert(m_PositionBound, this);
             m_CurrentPosition = context.ReadValue<Vector3>();
         }
 
-        void OnRotationUpdate(InputAction.CallbackContext context)
+        void OnRotationPerformed(InputAction.CallbackContext context)
         {
-            Debug.Assert(m_RotationBound);
+            Debug.Assert(m_RotationBound, this);
             m_CurrentRotation = context.ReadValue<Quaternion>();
         }
 
+        /// <summary>
+        /// This function is called when the script instance is being loaded.
+        /// </summary>
         protected virtual void Awake()
         {
 #if UNITY_INPUT_SYSTEM_ENABLE_VR && ENABLE_VR
@@ -155,18 +273,27 @@ namespace UnityEngine.InputSystem.XR
 #endif
         }
 
+        /// <summary>
+        /// This function is called when the object becomes enabled and active.
+        /// </summary>
         protected void OnEnable()
         {
             InputSystem.onAfterUpdate += UpdateCallback;
             BindActions();
         }
 
-        void OnDisable()
+        /// <summary>
+        /// This function is called when the object becomes disabled or inactive.
+        /// </summary>
+        protected void OnDisable()
         {
             UnbindActions();
             InputSystem.onAfterUpdate -= UpdateCallback;
         }
 
+        /// <summary>
+        /// This function is called when the <see cref="MonoBehaviour"/> will be destroyed.
+        /// </summary>
         protected virtual void OnDestroy()
         {
 #if UNITY_INPUT_SYSTEM_ENABLE_VR && ENABLE_VR
@@ -218,15 +345,31 @@ namespace UnityEngine.InputSystem.XR
             }
         }
 
-        private bool HasStereoCamera()
+        bool HasStereoCamera()
         {
-            var camera = GetComponent<Camera>();
-            return camera != null && camera.stereoEnabled;
+            var cameraComponent = GetComponent<Camera>();
+            return cameraComponent != null && cameraComponent.stereoEnabled;
         }
 
         protected virtual void PerformUpdate()
         {
             SetLocalTransform(m_CurrentPosition, m_CurrentRotation);
+        }
+
+        /// <inheritdoc />
+        void ISerializationCallbackReceiver.OnBeforeSerialize()
+        {
+        }
+
+        /// <inheritdoc />
+        void ISerializationCallbackReceiver.OnAfterDeserialize()
+        {
+            if (m_HasMigratedActions)
+                return;
+
+            m_PositionInput = new InputActionProperty(m_PositionAction);
+            m_RotationInput = new InputActionProperty(m_RotationAction);
+            m_HasMigratedActions = true;
         }
     }
 }